requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.9.2
SQLAlchemy==2.0.41
Werkzeug==3.1.3
WTForms==3.2.1
openpyxl==3.1.5
anthropic==0.49.0
msal==1.32.3
requests==2.32.3
bcrypt==4.3.0
bleach==6.2.0
python-dotenv==1.1.0
spacy>=3.7,<4.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
pytest==8.3.5
pytest-flask==1.3.0
